+++ dojo/test/__init__.py
@@ -1,0 +1,1 @@
+import dojo.test.admin  # noqa: F401
